refactor: separate authentication account from commands and queries
