+++ src/application/usecases/account/commands/index.ts
@@ -1,4 +1,2 @@
-export * from './db-add-account'
-export * from './db-authentication'
 export * from './db-change-account-email'
 export * from './db-change-account-password'
