+++ src/domain/usecases/account/commands/index.ts
@@ -1,4 +1,2 @@
-export * from './add-account'
-export * from './authentication'
 export * from './change-account-email'
 export * from './change-account-password'
