+++ src/main/factories/controllers/account/authentication/sign-in-controller-factory.ts
@@ -1,4 +1,4 @@
-import { AuthenticationFactory } from '@/main/factories/usecases/account/commands'
+import { AuthenticationFactory } from '@/main/factories/usecases/account/authentication'
 import { LogErrorControllerDecoratorFactory } from '@/main/factories/decorators/log'
 import { type Controller } from '@/presentation/protocols'
 import { SignInController } from '@/presentation/controllers/account/authentication'
