+++ src/main/factories/usecases/account/authentication/index.ts
@@ -1,0 +1,2 @@
+export * from './add-account-factory'
+export * from './authentication-factory'
