+++ src/main/factories/usecases/account/commands/index.ts
@@ -1,4 +1,2 @@
-export * from './add-account-factory'
-export * from './authentication-factory'
 export * from './change-account-email-factory'
 export * from './change-account-password-factory'
