+++ tests/presentation/mocks/account/index.ts
@@ -1,1 +1,3 @@
-export * from './mock-account'
+export * from './mock-account-authentication'
+export * from './mock-account-commands'
+export * from './mock-account-queries'
